+++ packages/spacecat-shared-content-client/CHANGELOG.md
@@ -1,3 +1,22 @@
+## [@adobe/spacecat-shared-content-client-v1.9.3](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.9.2...@adobe/spacecat-shared-content-client-v1.9.3) (2026-05-30)
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#1641](https://github.com/adobe/spacecat-shared/issues/1641)) ([aa28a98](https://github.com/adobe/spacecat-shared/commit/aa28a9875a610fcf96efa693ba652603e12053de))
+
+## [@adobe/spacecat-shared-content-client-v1.9.2](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.9.1...@adobe/spacecat-shared-content-client-v1.9.2) (2026-05-29)
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#1609](https://github.com/adobe/spacecat-shared/issues/1609)) ([ccdcc2b](https://github.com/adobe/spacecat-shared/commit/ccdcc2b88c1b03e5c58ed682daaaadb382638395))
+* **deps:** update external major (major) ([#1611](https://github.com/adobe/spacecat-shared/issues/1611)) ([5a80797](https://github.com/adobe/spacecat-shared/commit/5a807979187f1ebcc32b088faee9165aa38cc461))
+
+## [@adobe/spacecat-shared-content-client-v1.9.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.9.0...@adobe/spacecat-shared-content-client-v1.9.1) (2026-05-23)
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#1623](https://github.com/adobe/spacecat-shared/issues/1623)) ([19cb2a4](https://github.com/adobe/spacecat-shared/commit/19cb2a4fec00735ad4eead3030f9ea6b2573fdc3))
+
 ## [@adobe/spacecat-shared-content-client-v1.9.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.8.24...@adobe/spacecat-shared-content-client-v1.9.0) (2026-05-21)
 
 ### Features