+++ packages/spacecat-shared-utils/CHANGELOG.md
@@ -1,3 +1,27 @@
+## [@adobe/spacecat-shared-utils-v1.116.5](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.116.4...@adobe/spacecat-shared-utils-v1.116.5) (2026-05-30)
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#1641](https://github.com/adobe/spacecat-shared/issues/1641)) ([aa28a98](https://github.com/adobe/spacecat-shared/commit/aa28a9875a610fcf96efa693ba652603e12053de))
+
+## [@adobe/spacecat-shared-utils-v1.116.4](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.116.3...@adobe/spacecat-shared-utils-v1.116.4) (2026-05-29)
+
+### Bug Fixes
+
+* **deps:** update external major (major) ([#1611](https://github.com/adobe/spacecat-shared/issues/1611)) ([5a80797](https://github.com/adobe/spacecat-shared/commit/5a807979187f1ebcc32b088faee9165aa38cc461))
+
+## [@adobe/spacecat-shared-utils-v1.116.3](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.116.2...@adobe/spacecat-shared-utils-v1.116.3) (2026-05-27)
+
+### Bug Fixes
+
+* **data-access:** mint UUID v7 ids to match DB schema default ([#1633](https://github.com/adobe/spacecat-shared/issues/1633)) ([5c8f6fc](https://github.com/adobe/spacecat-shared/commit/5c8f6fcd454975703b4f670c0f560f43a768736a))
+
+## [@adobe/spacecat-shared-utils-v1.116.2](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.116.1...@adobe/spacecat-shared-utils-v1.116.2) (2026-05-23)
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#1623](https://github.com/adobe/spacecat-shared/issues/1623)) ([19cb2a4](https://github.com/adobe/spacecat-shared/commit/19cb2a4fec00735ad4eead3030f9ea6b2573fdc3))
+
 ## [@adobe/spacecat-shared-utils-v1.116.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.116.0...@adobe/spacecat-shared-utils-v1.116.1) (2026-05-22)
 
 ### Bug Fixes